Fix pin-rebuild-pin loop
This patch should fix the pin-rebuild-pin loop, that appears due to each
change in the repository resulting in a new image build, which results
in a new pin in the docker-compose file, which results in a merge, that
triggers a new build.

Instead this patch excludes the `docker-compose.yaml` from pinning and
solves the issue.

Signed-off-by: Christoph Kern <sheogorath@shivering-isles.com>